Split apart text.rs (#1350)
Fairly dumb refactoring, just splitting a big file down into more
smaller files with clearer inputs
