+++ qlty-cli/src/ui.rs
@@ -1,6 +1,16 @@
-pub mod format;
+mod errors;
+mod fixes;
 mod highlight;
+mod invocations;
+mod issues;
+mod level;
+mod source;
 mod steps;
+mod text;
+mod unformatted;
 
+pub use errors::ErrorsFormatter;
+pub use fixes::ApplyMode;
 pub use highlight::Highlighter;
 pub use steps::Steps;
+pub use text::TextFormatter;
