+++ qlty-cli/src/ui/errors.rs
