--- qlty-cli/src/ui/format.rs
@@ -1,5 +1,0 @@
-mod errors;
-mod text;
-
-pub use errors::ErrorsFormatter;
-pub use text::{ApplyMode, TextFormatter};
